# apps/api/src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: Memes API
  version: "2.0.0"
  description: |
    REST API for browsing and searching scraped Reddit memes.
    Push changes to memes/ to trigger automatic re-indexing via GitHub Actions.

    V2: Multilingual schema. Every read endpoint accepts a `?locale=` query
    parameter. The `Meme` payload carries a `translations[]` array (one entry
    per locale) and an `images[]` array (multiple images per meme, with
    `is_primary` marking the lead image).

servers:
  - url: /

tags:
  - name: memes
    description: Read-only browsing and search
  - name: admin
    description: Owner-only operations (requires X-Api-Key header)

paths:

  /:
    get:
      tags: [memes]
      operationId: getStats
      summary: Global collection statistics
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /categories:
    get:
      tags: [memes]
      operationId: listCategories
      summary: Paginated list of categories with meme counts and top score
      security:
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/LocaleParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /memes:
    get:
      tags: [memes]
      operationId: listMemes
      summary: Paginated list of memes with optional filters
      security:
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [score, created_at, title]
            default: score
        - $ref: "#/components/parameters/LocaleParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /memes/{category}:
    get:
      tags: [memes]
      operationId: listMemesByCategory
      summary: All memes in a specific category
      security:
        - ApiKeyAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [score, created_at, title]
            default: score
        - $ref: "#/components/parameters/LocaleParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /memes/{category}/{slug}:
    get:
      tags: [memes]
      operationId: getMeme
      summary: Single meme by category and slug
      security:
        - ApiKeyAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/LocaleParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meme"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Meme not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /search:
    get:
      tags: [memes]
      operationId: searchMemes
      summary: Locale-aware full-text search across title and description
      security:
        - ApiKeyAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - $ref: "#/components/parameters/LocaleParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
        "400":
          description: Missing or empty q parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /admin/reindex:
    post:
      tags: [admin]
      operationId: reindex
      summary: Upsert a single meme (with body) or re-scan the full memes/ and _categories/ directories (no body)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: indexMemes
          in: query
          description: Whether to index memes from memes/ directory (default true)
          required: false
          schema:
            type: boolean
            default: true
        - name: indexCategories
          in: query
          description: Whether to index categories from memes/_categories/ directory (default true)
          required: false
          schema:
            type: boolean
            default: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemeIndexRequest"
      responses:
        "200":
          description: Reindex accepted and processing asynchronously
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReindexAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /admin/api-keys:
    get:
      tags: [admin]
      operationId: listApiKeys
      summary: List all active API keys
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      tags: [admin]
      operationId: createApiKey
      summary: Create a new API key (returned once in plaintext)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /admin/api-keys/{id}:
    delete:
      tags: [admin]
      operationId: revokeApiKey
      summary: Revoke an API key
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key

  parameters:
    LocaleParam:
      name: locale
      in: query
      description: Locale for translatable fields. Defaults to `en`.
      schema:
        $ref: "#/components/schemas/LocaleCode"

  responses:
    Unauthorized:
      description: Missing or invalid X-Api-Key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions for this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    LocaleCode:
      type: string
      enum: [en, es, pt, fr, de, ar, es-ar]
      default: en

    Stats:
      type: object
      properties:
        total_memes:
          type: integer
        total_categories:
          type: integer
        total_subreddits:
          type: integer
        top_category:
          type: string
        indexed_at:
          type: string
          format: date-time

    CategoryTranslation:
      type: object
      required: [locale, name]
      properties:
        locale:
          $ref: "#/components/schemas/LocaleCode"
        name:
          type: string
        description:
          type: string
          nullable: true

    CategorySummary:
      type: object
      required: [category, count, top_score]
      properties:
        category:
          type: string
        count:
          type: integer
        top_score:
          type: integer
        translations:
          type: array
          items:
            $ref: "#/components/schemas/CategoryTranslation"
        images:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/CategoryImage"

    MemeTranslation:
      type: object
      required: [locale, title]
      properties:
        locale:
          $ref: "#/components/schemas/LocaleCode"
        title:
          type: string
        description:
          type: string
          nullable: true

    MemeImage:
      type: object
      required: [path, position, is_primary]
      properties:
        path:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        bytes:
          type: integer
          format: int64
          nullable: true
        mime_type:
          type: string
          nullable: true
        position:
          type: integer
        is_primary:
          type: boolean

    CategoryImage:
      type: object
      required: [path, image_type, position, is_primary]
      properties:
        path:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        bytes:
          type: integer
          format: int64
          nullable: true
        mime_type:
          type: string
          nullable: true
        image_type:
          type: string
          enum: [icon, banner, thumbnail]
        position:
          type: integer
        is_primary:
          type: boolean

    Meme:
      type: object
      required: [slug, category, default_locale, translations, images]
      properties:
        slug:
          type: string
        category:
          type: string
        default_locale:
          $ref: "#/components/schemas/LocaleCode"
        author:
          type: string
          nullable: true
        subreddit:
          type: string
          nullable: true
        score:
          type: integer
        created_at:
          type: string
          format: date-time
          nullable: true
        source_url:
          type: string
          nullable: true
        post_url:
          type: string
          nullable: true
        translations:
          type: array
          items:
            $ref: "#/components/schemas/MemeTranslation"
        images:
          type: array
          items:
            $ref: "#/components/schemas/MemeImage"
        tags:
          type: array
          items:
            type: string

    SearchResult:
      type: object
      required: [slug, category, score, title, tags]
      properties:
        slug:
          type: string
        category:
          type: string
        author:
          type: string
          nullable: true
        score:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        image_path:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string

    MemeListItem:
      type: object
      required: [slug, category, score, title, tags]
      properties:
        slug:
          type: string
        score:
          type: integer
        created_at:
          type: string
          format: date-time
          nullable: true
        category:
          type: string
        author:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        image_path:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string

    CategoryPage:
      type: object
      required: [data, page, limit, total, total_pages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    MemePage:
      type: object
      required: [data, page, limit, total, total_pages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MemeListItem"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    ReindexResult:
      type: object
      properties:
        indexed:
          type: integer
        duration_ms:
          type: integer
          format: int64
        errors:
          type: array
          items:
            type: string

    MemeIndexRequest:
      type: object
      required: [slug, category, translations]
      properties:
        slug:
          type: string
        category:
          type: string
        default_locale:
          $ref: "#/components/schemas/LocaleCode"
        author:
          type: string
        subreddit:
          type: string
        score:
          type: integer
          format: int32
          default: 0
        source_url:
          type: string
        post_url:
          type: string
        created_at:
          type: string
          format: date-time
        translations:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/MemeTranslation"
        images:
          type: array
          items:
            $ref: "#/components/schemas/MemeImage"
        tags:
          type: array
          items:
            type: string

    ReindexAccepted:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted]
          example: accepted

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

    ApiKey:
      type: object
      properties:
        id:
          type: integer
          format: int64
        client_name:
          type: string
        role:
          type: string
          enum: [READ, WRITE, ADMIN]
        active:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true

    ApiKeyCreateRequest:
      type: object
      required: [client_name, role]
      properties:
        client_name:
          type: string
        role:
          type: string
          enum: [READ, WRITE, ADMIN]
        expires_at:
          type: string
          format: date-time
          nullable: true

    ApiKeyCreated:
      type: object
      required: [key, id]
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
          description: "The plaintext API key. Store securely — it will never be shown again."